FIX FINAL PRAKTIKUM

+++ app/Http/Controllers/StudentController.php
@@ -2,32 +2,45 @@
 
 namespace App\Http\Controllers;
 
-use App\Models\Student;
 use Illuminate\Http\Request;
+use App\Models\Student;
 use App\Models\Majors;
+use Illuminate\Support\Facades\Gate;
 
 class StudentController extends Controller
 {
     public function index()
     {
+        if (! Gate::allows('view-student')) {
+            abort(401);
+        }
         $students = Student::with('majors')->get();
         return view('students.index', compact('students'));
     }
 
     public function show(string $id)
     {
+        if (! Gate::allows('view-student')) {
+            abort(401);
+        }
         $student = Student::with('majors')->find($id);
         return view('students.show', compact('student'));
     }
 
     public function create()
     {
+        if (! Gate::allows('store-student')) {
+            abort(401);
+        }
         $majors = Majors::all();
         return view('students.create', compact('majors'));
     }
 
     public function store(Request $request)
     {
+        if (! Gate::allows('store-student')) {
+            abort(401);
+        }
         $validated = $request->validate([
             'name' => 'required',
             'student_id_number' => 'required|unique:students|max:9',
@@ -53,15 +66,23 @@ class StudentController extends Controller
         return redirect()->route('students.index')->with('success', 'Student created successfully');
     }
 
+    // Menampilkan form edit student
     public function edit(string $id)
     {
+        if (! Gate::allows('edit-student')) {
+            abort(401);
+        }
         $student = Student::with('majors')->find($id);
         $majors = Majors::all();
+
         return view('students.edit', compact('student', 'majors'));
     }
 
     public function update(Request $request, string $id)
     {
+        if (! Gate::allows('edit-student')) {
+            abort(401);
+        }
         $validated = $request->validate([
             'name' => 'required',
             'student_id_number' => "required|unique:students,student_id_number,$id|max:9",
@@ -88,9 +109,11 @@ class StudentController extends Controller
 
         return redirect()->route('students.index')->with('success', 'Student updated successfully');
     }
-
     public function destroy(string $id)
     {
+        if (! Gate::allows('destroy-student')) {
+            abort(401);
+        }
         $student = Student::findOrFail($id);
         $student->delete();
